Move test for page metadata
